requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
fake-useragent==0.1.11
idna==3.2
lxml==4.6.3
numpy==1.21.1
opencv-python==4.5.3.56
peewee==3.14.4
Pillow==8.3.1
psycopg2==2.9.1
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
